refactor(label): refactor label layout utils
